refactor(core): using Context to store our isKeyboadUser mode
