+++ components/core/src/index.js
@@ -1,4 +1,4 @@
 export { jsx, css, Global, ClassNames } from '@emotion/core';
-export { GEL, ThemeContext, useTheme } from './Theme';
+export { GEL, ThemeContext, useTheme, useIsKeyboardUser } from './Theme';
 export { Core } from './Core';
 export { paint } from './utils';
